refactor(services): extract auth service and server rate limiter
Move auth implementation into src/services/auth, introduce shared RateLimiter, and keep src/auth.ts as compatibility re-export.

Refs: #1614
