+++ src/server.ts
@@ -41,7 +41,7 @@ import { SSEWriter } from './sse-writer.js';
 import { SSEConnectionLimiter } from './sse-limiter.js';
 import { PipelineManager } from './pipeline.js';
 import { ToolRegistry } from './tool-registry.js';
-import { AuthManager, classifyBearerTokenForRoute, type ApiKeyRole } from './auth.js';
+import { AuthManager, RateLimiter, classifyBearerTokenForRoute, type ApiKeyRole } from './services/auth/index.js';
 import { AuditLogger, type AuditAction } from './audit.js';
 import { MetricsCollector } from './metrics.js';
 import { promRegistry, METRICS_CONTENT_TYPE } from './prometheus.js';
@@ -274,107 +274,7 @@ app.addHook('onSend', (req, reply, payload, done) => {
 });
 
 // Auth middleware setup (Issue #39: multi-key auth with rate limiting)
-// #228: Per-IP rate limiting (applies even with master token, with higher limits)
-// #622: Circular buffer — O(1) prune via index advancement instead of O(n) shift()
-interface IpRateBucket {
-  entries: number[];
-  start: number;
-}
-const ipRateLimits = new Map<string, IpRateBucket>();
-const IP_WINDOW_MS = 60_000;
-const IP_LIMIT_NORMAL = 120;   // per minute for regular keys
-const IP_LIMIT_MASTER = 300;   // per minute for master token
-const MAX_IP_ENTRIES = 10_000; // #844: Cap tracked IPs to prevent memory exhaustion
-
-function checkIpRateLimit(ip: string, isMaster: boolean): boolean {
-  const now = Date.now();
-  const cutoff = now - IP_WINDOW_MS;
-  const bucket = ipRateLimits.get(ip) || { entries: [], start: 0 };
-  // O(1) prune: advance start index past expired entries
-  while (bucket.start < bucket.entries.length && bucket.entries[bucket.start]! < cutoff) {
-    bucket.start++;
-  }
-  // Compact when the leading garbage exceeds 50% of the allocated array
-  if (bucket.start > bucket.entries.length >>> 1) {
-    bucket.entries = bucket.entries.slice(bucket.start);
-    bucket.start = 0;
-  }
-  bucket.entries.push(now);
-  ipRateLimits.set(ip, bucket);
-  // #844: Evict oldest IPs when map exceeds cap to prevent unbounded memory growth
-  if (ipRateLimits.size > MAX_IP_ENTRIES) {
-    let oldestIp = '';
-    let oldestTime = Infinity;
-    for (const [trackedIp, trackedBucket] of ipRateLimits) {
-      const lastTs = trackedBucket.entries[trackedBucket.entries.length - 1];
-      if (lastTs !== undefined && lastTs < oldestTime) {
-        oldestTime = lastTs;
-        oldestIp = trackedIp;
-      }
-    }
-    if (oldestIp) ipRateLimits.delete(oldestIp);
-  }
-  const activeCount = bucket.entries.length - bucket.start;
-  const limit = isMaster ? IP_LIMIT_MASTER : IP_LIMIT_NORMAL;
-  return activeCount > limit;
-}
-
-// #632: Auth failure rate limiting — 5 failed auth attempts per minute per IP.
-interface AuthFailBucket {
-  timestamps: number[];
-}
-const authFailLimits = new Map<string, AuthFailBucket>();
-const AUTH_FAIL_WINDOW_MS = 60_000;
-const AUTH_FAIL_MAX = 5;
-const MAX_AUTH_FAIL_IP_ENTRIES = 10_000;
-
-function checkAuthFailRateLimit(ip: string): boolean {
-  const now = Date.now();
-  const cutoff = now - AUTH_FAIL_WINDOW_MS;
-  const bucket = authFailLimits.get(ip) || { timestamps: [] };
-  // Prune expired entries
-  bucket.timestamps = bucket.timestamps.filter(t => t >= cutoff);
-  bucket.timestamps.push(now);
-  authFailLimits.set(ip, bucket);
-  if (authFailLimits.size > MAX_AUTH_FAIL_IP_ENTRIES) {
-    let oldestIp = '';
-    let oldestTime = Infinity;
-    for (const [trackedIp, trackedBucket] of authFailLimits) {
-      const lastTs = trackedBucket.timestamps[trackedBucket.timestamps.length - 1];
-      if (lastTs !== undefined && lastTs < oldestTime) {
-        oldestTime = lastTs;
-        oldestIp = trackedIp;
-      }
-    }
-    if (oldestIp) authFailLimits.delete(oldestIp);
-  }
-  return bucket.timestamps.length > AUTH_FAIL_MAX;
-}
-
-function recordAuthFailure(ip: string): void {
-  checkAuthFailRateLimit(ip);
-}
-
-/** #632: Prune stale auth-failure buckets. */
-function pruneAuthFailLimits(): void {
-  const cutoff = Date.now() - AUTH_FAIL_WINDOW_MS;
-  for (const [ip, bucket] of authFailLimits) {
-    bucket.timestamps = bucket.timestamps.filter(t => t >= cutoff);
-    if (bucket.timestamps.length === 0) authFailLimits.delete(ip);
-  }
-}
-
-/** #357: Prune IPs whose timestamp arrays are entirely outside the rate-limit window. */
-function pruneIpRateLimits(): void {
-  const cutoff = Date.now() - IP_WINDOW_MS;
-  for (const [ip, bucket] of ipRateLimits) {
-    // All timestamps are old — remove the entry entirely
-    const last = bucket.entries[bucket.entries.length - 1];
-    if (bucket.entries.length - bucket.start === 0 || (last !== undefined && last < cutoff)) {
-      ipRateLimits.delete(ip);
-    }
-  }
-}
+const rateLimiter = new RateLimiter();
 
 /** #583: Track keyId per request for batch rate limiting. */
 const requestKeyMap = new Map<string, string>();
@@ -472,7 +372,7 @@ function setupAuth(authManager: AuthManager): void {
     // #633: Only use req.ip — trustProxy controls whether X-Forwarded-For is considered
     const clientIp = req.ip ?? 'unknown';
     // #632: Block IPs that exceeded auth failure rate limit (5 attempts/min)
-    if (checkAuthFailRateLimit(clientIp)) {
+    if (rateLimiter.checkAuthFailRateLimit(clientIp)) {
       return reply.status(429).send({ error: 'Too many auth failures — try again later' });
     }
 
@@ -484,19 +384,19 @@ function setupAuth(authManager: AuthManager): void {
       if (await authManager.validateSSEToken(token)) {
         return; // authenticated via short-lived SSE token
       }
-      recordAuthFailure(clientIp);
+      rateLimiter.recordAuthFailure(clientIp);
       return reply.status(401).send({ error: 'Unauthorized — SSE token invalid or expired' });
     }
 
     if (tokenMode === 'reject') {
-      recordAuthFailure(clientIp);
+      rateLimiter.recordAuthFailure(clientIp);
       return reply.status(401).send({ error: 'Unauthorized — SSE token required for event streams' });
     }
 
     const result = authManager.validate(token);
 
     if (!result.valid) {
-      recordAuthFailure(clientIp);
+      rateLimiter.recordAuthFailure(clientIp);
       // Issue #1403: Distinguish expired keys from invalid keys
       if (result.reason === 'expired') {
         return reply.status(401).send({ error: 'Unauthorized — API key has expired', code: 'KEY_EXPIRED' });
@@ -521,7 +421,7 @@ function setupAuth(authManager: AuthManager): void {
     // #228: Per-IP rate limiting (applies to all authenticated requests)
     // #633: Only use req.ip — trustProxy controls whether X-Forwarded-For is considered
     const isMaster = result.keyId === 'master';
-    if (checkIpRateLimit(clientIp, isMaster)) {
+    if (rateLimiter.checkIpRateLimit(clientIp, isMaster)) {
       return reply.status(429).send({ error: 'Rate limit exceeded — IP throttled' });
     }
   });
@@ -641,7 +541,7 @@ app.post('/v1/auth/verify', async (req, reply) => {
 
   // Public bootstrap endpoint: apply failed-auth IP throttling like the main auth hook.
   const clientIp = req.ip ?? 'unknown';
-  if (checkAuthFailRateLimit(clientIp)) {
+  if (rateLimiter.checkAuthFailRateLimit(clientIp)) {
     return reply.status(429).send({ valid: false });
   }
 
@@ -650,7 +550,7 @@ app.post('/v1/auth/verify', async (req, reply) => {
     return reply.status(429).send({ valid: false });
   }
   if (!result.valid) {
-    recordAuthFailure(clientIp);
+    rateLimiter.recordAuthFailure(clientIp);
     return reply.status(401).send({ valid: false });
   }
 
@@ -2265,10 +2165,9 @@ async function main(): Promise<void> {
   const reaperInterval = setInterval(() => reapStaleSessions(config.maxSessionAgeMs), config.reaperIntervalMs);
   const zombieReaperInterval = setInterval(() => reapZombieSessions(), ZOMBIE_REAP_INTERVAL_MS);
   const metricsSaveInterval = setInterval(() => { void metrics.save(); }, 5 * 60 * 1000);
-  // #357: Prune stale IP rate-limit entries every minute
-  const ipPruneInterval = setInterval(pruneIpRateLimits, 60_000);
-  // #632: Prune stale auth failure rate-limit buckets every minute
-  const authFailPruneInterval = setInterval(pruneAuthFailLimits, 60_000);
+  // #357/#632: Prune stale rate-limit entries every minute
+  const ipPruneInterval = setInterval(() => rateLimiter.pruneIpRateLimits(), 60_000);
+  const authFailPruneInterval = setInterval(() => rateLimiter.pruneAuthFailLimits(), 60_000);
   // #398: Sweep stale API key rate limit buckets every 5 minutes
   const authSweepInterval = setInterval(() => auth.sweepStaleRateLimits(), 5 * 60_000);
   let pidFilePath = '';
