+++ src/services/auth/index.ts
@@ -1,0 +1,3 @@
+export { AuthManager, classifyBearerTokenForRoute } from './AuthManager.js';
+export { RateLimiter } from './RateLimiter.js';
+export type { ApiKey, ApiKeyRole, ApiKeyStore, AuthRejectReason } from './types.js';
